requirements.txt
streamlit
gspread
pandas
gTTS
google-auth
